+++ src/android/FirebasePlugin.java
@@ -1,17 +1,23 @@
 package org.apache.cordova.firebase;
 
+import android.app.Activity;
 import android.app.NotificationManager;
-import android.content.ComponentName;
+import android.app.NotificationChannel;
+import android.content.ContentResolver;
+import android.media.RingtoneManager;
+import android.net.Uri;
+import android.media.AudioAttributes;
 import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
+import android.os.Build;
 import android.os.Bundle;
+import androidx.core.app.NotificationCompat;
 import androidx.core.app.NotificationManagerCompat;
 import android.util.Base64;
 import android.util.Log;
 
 import com.crashlytics.android.Crashlytics;
-import io.fabric.sdk.android.Fabric;
 import java.lang.reflect.Field;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
@@ -30,9 +36,9 @@ import com.google.firebase.perf.metrics.Trace;
 import me.leolin.shortcutbadger.ShortcutBadger;
 
 import org.apache.cordova.CallbackContext;
+import org.apache.cordova.CordovaInterface;
 import org.apache.cordova.CordovaPlugin;
 import org.apache.cordova.PluginResult;
-import org.apache.cordova.CordovaWebView;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -42,14 +48,12 @@ import java.util.HashMap;
 import java.util.Iterator;
 import java.util.Map;
 import java.util.Set;
+import java.util.List;
 
 // Firebase PhoneAuth
 import java.util.concurrent.TimeUnit;
 
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.AuthResult;
 import com.google.firebase.FirebaseException;
-import com.google.firebase.auth.FirebaseAuthException;
 import com.google.firebase.auth.FirebaseAuthInvalidCredentialsException;
 import com.google.firebase.FirebaseTooManyRequestsException;
 import com.google.firebase.auth.PhoneAuthCredential;
@@ -58,8 +62,10 @@ import com.google.firebase.auth.PhoneAuthProvider;
 public class FirebasePlugin extends CordovaPlugin {
 
     private FirebaseAnalytics mFirebaseAnalytics;
-    private static CordovaWebView appView;
-    private final String TAG = "FirebasePlugin";
+    private static CordovaInterface cordovaInterface = null;
+    private static Context applicationContext = null;
+    private static Activity cordovaActivity;
+    private static final String TAG = "FirebasePlugin";
     protected static final String KEY = "badge";
 
     private static boolean inBackground = true;
@@ -67,24 +73,38 @@ public class FirebasePlugin extends CordovaPlugin {
     private static CallbackContext notificationCallbackContext;
     private static CallbackContext tokenRefreshCallbackContext;
 
+    private static NotificationChannel defaultNotificationChannel = null;
+    public static String defaultChannelId = null;
+    public static String defaultChannelName = null;
+
     @Override
     protected void pluginInitialize() {
-        final Context context = this.cordova.getActivity().getApplicationContext();
-        final Bundle extras = this.cordova.getActivity().getIntent().getExtras();
+        cordovaActivity = this.cordova.getActivity();
+        applicationContext = cordovaActivity.getApplicationContext();
+        final Bundle extras = cordovaActivity.getIntent().getExtras();
+        FirebasePlugin.cordovaInterface = this.cordova;
         this.cordova.getThreadPool().execute(new Runnable() {
             public void run() {
-                Log.d(TAG, "Starting Firebase plugin");
-                FirebaseApp.initializeApp(context);
-                mFirebaseAnalytics = FirebaseAnalytics.getInstance(context);
-                mFirebaseAnalytics.setAnalyticsCollectionEnabled(true);
-                if (extras != null && extras.size() > 1) {
-                    if (FirebasePlugin.notificationStack == null) {
-                        FirebasePlugin.notificationStack = new ArrayList<Bundle>();
+                try {
+                    Log.d(TAG, "Starting Firebase plugin");
+                    FirebaseApp.initializeApp(applicationContext);
+                    mFirebaseAnalytics = FirebaseAnalytics.getInstance(applicationContext);
+                    mFirebaseAnalytics.setAnalyticsCollectionEnabled(true);
+                    if (extras != null && extras.size() > 1) {
+                        if (FirebasePlugin.notificationStack == null) {
+                            FirebasePlugin.notificationStack = new ArrayList<Bundle>();
+                        }
+                        if (extras.containsKey("google.message_id")) {
+                            extras.putBoolean("tap", true);
+                            notificationStack.add(extras);
+                        }
                     }
-                    if (extras.containsKey("google.message_id")) {
-                        extras.putBoolean("tap", true);
-                        notificationStack.add(extras);
-                    }
+                    defaultChannelId = getStringResource("default_notification_channel_id");
+                    defaultChannelName = getStringResource("default_notification_channel_name");
+                    createDefaultChannel();
+                }catch (Exception e){
+                    Log.e(TAG, e.getMessage());
+                    Crashlytics.logException(e);
                 }
             }
         });
@@ -119,8 +139,8 @@ public class FirebasePlugin extends CordovaPlugin {
         } else if (action.equals("unregister")) {
             this.unregister(callbackContext);
             return true;
-        } else if (action.equals("onNotificationOpen")) {
-            this.onNotificationOpen(callbackContext);
+        } else if (action.equals("onMessageReceived")) {
+            this.onMessageReceived(callbackContext);
             return true;
         } else if (action.equals("onTokenRefresh")) {
             this.onTokenRefresh(callbackContext);
@@ -195,6 +215,21 @@ public class FirebasePlugin extends CordovaPlugin {
         } else if (action.equals("sendCrash")) {
             sendCrash(args, callbackContext);
             return true;
+        } else if (action.equals("createChannel")) {
+            this.createChannel(callbackContext, args.getJSONObject(0));
+            return true;
+        } else if (action.equals("deleteChannel")) {
+            this.deleteChannel(callbackContext, args.getString(0));
+            return true;
+        } else if (action.equals("listChannels")) {
+            this.listChannels(callbackContext);
+            return true;
+        } else if (action.equals("setDefaultChannel")) {
+            this.setDefaultChannel(callbackContext, args.getJSONObject(0));
+            return true;
+        } else if (action.equals("hasPermission") || action.equals("grantPermission")) {
+            callbackContext.sendPluginResult(new PluginResult(PluginResult.Status.OK, true));
+            return true;
         }
 
         return false;
@@ -219,17 +254,27 @@ public class FirebasePlugin extends CordovaPlugin {
     @Override
     public void onDestroy() {
         super.onDestroy();
-
-        if (this.appView != null) {
-            appView.handleDestroy();
-        }
     }
 
-    private void onNotificationOpen(final CallbackContext callbackContext) {
+    /**
+     * Get a string from resources without importing the .R package
+     *
+     * @param name Resource Name
+     * @return Resource
+     */
+    private String getStringResource(String name) {
+        return applicationContext.getString(
+                applicationContext.getResources().getIdentifier(
+                        name, "string", applicationContext.getPackageName()
+                )
+        );
+    }
+
+    private void onMessageReceived(final CallbackContext callbackContext) {
         FirebasePlugin.notificationCallbackContext = callbackContext;
         if (FirebasePlugin.notificationStack != null) {
             for (Bundle bundle : FirebasePlugin.notificationStack) {
-                FirebasePlugin.sendNotification(bundle, this.cordova.getActivity().getApplicationContext());
+                FirebasePlugin.sendMessage(bundle, cordovaActivity.getApplicationContext());
             }
             FirebasePlugin.notificationStack.clear();
         }
@@ -253,7 +298,7 @@ public class FirebasePlugin extends CordovaPlugin {
         });
     }
 
-    public static void sendNotification(Bundle bundle, Context context) {
+    public static void sendMessage(Bundle bundle, Context context) {
         if (!FirebasePlugin.hasNotificationsCallback()) {
             String packageName = context.getPackageName();
             if (FirebasePlugin.notificationStack == null) {
@@ -310,7 +355,7 @@ public class FirebasePlugin extends CordovaPlugin {
         final Bundle data = intent.getExtras();
         if (data != null && data.containsKey("google.message_id")) {
             data.putBoolean("tap", true);
-            FirebasePlugin.sendNotification(data, this.cordova.getActivity().getApplicationContext());
+            FirebasePlugin.sendMessage(data, applicationContext);
         }
     }
 
@@ -360,8 +405,7 @@ public class FirebasePlugin extends CordovaPlugin {
         cordova.getThreadPool().execute(new Runnable() {
             public void run() {
                 try {
-                    Context context = cordova.getActivity();
-                    NotificationManagerCompat notificationManagerCompat = NotificationManagerCompat.from(context);
+                    NotificationManagerCompat notificationManagerCompat = NotificationManagerCompat.from(cordovaActivity);
                     boolean areNotificationsEnabled = notificationManagerCompat.areNotificationsEnabled();
                     JSONObject object = new JSONObject();
                     object.put("isEnabled", areNotificationsEnabled);
@@ -378,11 +422,10 @@ public class FirebasePlugin extends CordovaPlugin {
         cordova.getThreadPool().execute(new Runnable() {
             public void run() {
                 try {
-                    Context context = cordova.getActivity();
-                    SharedPreferences.Editor editor = context.getSharedPreferences(KEY, Context.MODE_PRIVATE).edit();
+                    SharedPreferences.Editor editor = cordovaActivity.getSharedPreferences(KEY, Context.MODE_PRIVATE).edit();
                     editor.putInt(KEY, number);
                     editor.apply();
-                    ShortcutBadger.applyCount(context, number);
+                    ShortcutBadger.applyCount(cordovaActivity, number);
                     callbackContext.success();
                 } catch (Exception e) {
                     Crashlytics.logException(e);
@@ -396,8 +439,7 @@ public class FirebasePlugin extends CordovaPlugin {
         cordova.getThreadPool().execute(new Runnable() {
             public void run() {
                 try {
-                    Context context = cordova.getActivity();
-                    SharedPreferences settings = context.getSharedPreferences(KEY, Context.MODE_PRIVATE);
+                    SharedPreferences settings = cordovaActivity.getSharedPreferences(KEY, Context.MODE_PRIVATE);
                     int number = settings.getInt(KEY, 0);
                     callbackContext.success(number);
                 } catch (Exception e) {
@@ -483,9 +525,10 @@ public class FirebasePlugin extends CordovaPlugin {
             public void run() {
                 try {
                     Crashlytics.logException(new Exception(message));
+                    Log.e(TAG, message);
                     callbackContext.success(1);
                 } catch (Exception e) {
-                    Crashlytics.log(e.getMessage());
+                    Crashlytics.log(1, TAG, e.getMessage());
                     e.printStackTrace();
                     callbackContext.error(e.getMessage());
                 }
@@ -504,7 +547,7 @@ public class FirebasePlugin extends CordovaPlugin {
     private void sendCrash(final JSONArray data,
 						   final CallbackContext callbackContext) {
 
-		this.cordova.getActivity().runOnUiThread(new Runnable() {
+        cordovaActivity.runOnUiThread(new Runnable() {
 			@Override
 			public void run() {
 				throw new RuntimeException("This is a crash");
@@ -514,7 +557,7 @@ public class FirebasePlugin extends CordovaPlugin {
 
 
     private void setCrashlyticsUserId(final CallbackContext callbackContext, final String userId) {
-        cordova.getActivity().runOnUiThread(new Runnable() {
+        cordovaActivity.runOnUiThread(new Runnable() {
             public void run() {
                 try {
                     Crashlytics.setUserIdentifier(userId);
@@ -529,10 +572,10 @@ public class FirebasePlugin extends CordovaPlugin {
 
     private void setScreenName(final CallbackContext callbackContext, final String name) {
         // This must be called on the main thread
-        cordova.getActivity().runOnUiThread(new Runnable() {
+        cordovaActivity.runOnUiThread(new Runnable() {
             public void run() {
                 try {
-                    mFirebaseAnalytics.setCurrentScreen(cordova.getActivity(), name, null);
+                    mFirebaseAnalytics.setCurrentScreen(cordovaActivity, name, null);
                     callbackContext.success();
                 } catch (Exception e) {
                     Crashlytics.logException(e);
@@ -826,7 +869,7 @@ public class FirebasePlugin extends CordovaPlugin {
                     PhoneAuthProvider.getInstance().verifyPhoneNumber(number, // Phone number to verify
                             timeOutDuration, // Timeout duration
                             TimeUnit.SECONDS, // Unit of timeout
-                            cordova.getActivity(), // Activity (for callback binding)
+                            cordovaActivity, // Activity (for callback binding)
                             mCallbacks); // OnVerificationStateChangedCallbacks
                 } catch (Exception e) {
                     Crashlytics.logException(e);
@@ -967,8 +1010,7 @@ public class FirebasePlugin extends CordovaPlugin {
         cordova.getThreadPool().execute(new Runnable() {
             public void run() {
                 try {
-                    Context context = cordova.getActivity();
-                    NotificationManager nm = (NotificationManager) context.getSystemService(Context.NOTIFICATION_SERVICE);
+                    NotificationManager nm = (NotificationManager) cordovaActivity.getSystemService(Context.NOTIFICATION_SERVICE);
                     nm.cancelAll();
                     callbackContext.success();
                 } catch (Exception e) {
@@ -976,5 +1018,212 @@ public class FirebasePlugin extends CordovaPlugin {
                 }
             }
         });
+    }
+
+    public void createChannel(final CallbackContext callbackContext, final JSONObject options) {
+        cordova.getThreadPool().execute(new Runnable() {
+            public void run() {
+                try {
+                    createChannel(options);
+                    callbackContext.success();
+                } catch (Exception e) {
+                    Crashlytics.log(e.getMessage());
+                    callbackContext.error(e.getMessage());
+                }
+            }
+        });
+    }
+
+    protected static NotificationChannel createChannel(final JSONObject options) throws JSONException {
+        NotificationChannel channel = null;
+        // only call on Android O and above
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            String id = options.getString("id");
+            Log.i(TAG, "Creating channel id="+id);
+
+            if(channelExists(id)){
+                deleteChannel(id);
+            }
+
+            NotificationManager nm = (NotificationManager) cordovaActivity.getSystemService(Context.NOTIFICATION_SERVICE);
+            String packageName = cordovaActivity.getPackageName();
+
+            String name = options.optString("name", "");
+            Log.d(TAG, "Channel "+id+" - name="+name);
+
+            int importance = options.optInt("importance", NotificationManager.IMPORTANCE_HIGH);
+            Log.d(TAG, "Channel "+id+" - importance="+importance);
+
+            channel = new NotificationChannel(id,
+                    name,
+                    importance);
+
+            // Light
+            boolean light = options.optBoolean("light", true);
+            Log.d(TAG, "Channel "+id+" - light="+light);
+            channel.enableLights(light);
+
+            int lightColor = options.optInt("lightColor", -1);
+            if (lightColor != -1) {
+                Log.d(TAG, "Channel "+id+" - lightColor="+lightColor);
+                channel.setLightColor(lightColor);
+            }
+
+            // Visibility
+            int visibility = options.optInt("visibility", NotificationCompat.VISIBILITY_PUBLIC);
+            Log.d(TAG, "Channel "+id+" - visibility="+visibility);
+            channel.setLockscreenVisibility(visibility);
+
+            // Badge
+            boolean badge = options.optBoolean("badge", true);
+            Log.d(TAG, "Channel "+id+" - badge="+badge);
+            channel.setShowBadge(badge);
+
+            // Sound
+            String sound = options.optString("sound", "default");
+            AudioAttributes audioAttributes = new AudioAttributes.Builder()
+                    .setContentType(AudioAttributes.CONTENT_TYPE_SONIFICATION)
+                    .setUsage(AudioAttributes.USAGE_NOTIFICATION_RINGTONE).build();
+            if ("ringtone".equals(sound)) {
+                channel.setSound(RingtoneManager.getDefaultUri(RingtoneManager.TYPE_RINGTONE), audioAttributes);
+                Log.d(TAG, "Channel "+id+" - sound=ringtone");
+            } else if (sound != null && !sound.contentEquals("default")) {
+                Uri soundUri = Uri.parse(ContentResolver.SCHEME_ANDROID_RESOURCE + "://" + packageName + "/raw/" + sound);
+                channel.setSound(soundUri, audioAttributes);
+                Log.d(TAG, "Channel "+id+" - sound="+sound);
+            } else if (sound != "false"){
+                channel.setSound(RingtoneManager.getDefaultUri(RingtoneManager.TYPE_NOTIFICATION), audioAttributes);
+                Log.d(TAG, "Channel "+id+" - sound=default");
+            }else{
+                Log.d(TAG, "Channel "+id+" - sound=none");
+            }
+
+            // Vibration: if vibration setting is an array set vibration pattern, else set enable vibration.
+            JSONArray pattern = options.optJSONArray("vibration");
+            if (pattern != null) {
+                int patternLength = pattern.length();
+                long[] patternArray = new long[patternLength];
+                for (int i = 0; i < patternLength; i++) {
+                    patternArray[i] = pattern.optLong(i);
+                }
+                channel.enableVibration(true);
+                channel.setVibrationPattern(patternArray);
+                Log.d(TAG, "Channel "+id+" - vibrate="+pattern);
+            } else {
+                boolean vibrate = options.optBoolean("vibration", true);
+                channel.enableVibration(vibrate);
+                Log.d(TAG, "Channel "+id+" - vibrate="+vibrate);
+            }
+
+            // Create channel
+            nm.createNotificationChannel(channel);
+        }
+        return channel;
+    }
+
+    protected static void createDefaultChannel() throws JSONException {
+        JSONObject options = new JSONObject();
+        options.put("id", defaultChannelId);
+        options.put("name", defaultChannelName);
+        createDefaultChannel(options);
+    }
+
+    protected static void createDefaultChannel(final JSONObject options) throws JSONException {
+        defaultNotificationChannel = createChannel(options);
+    }
+
+    public void setDefaultChannel(final CallbackContext callbackContext, final JSONObject options) {
+        cordova.getThreadPool().execute(new Runnable() {
+            public void run() {
+                try {
+                    deleteChannel(defaultChannelId);
+
+                    String id = options.optString("id", null);
+                    if(id != null){
+                        defaultChannelId = id;
+                    }
+
+                    String name = options.optString("name", null);
+                    if(name != null){
+                        defaultChannelName = name;
+                    }
+                    createDefaultChannel(options);
+                    callbackContext.success();
+                } catch (Exception e) {
+                    Crashlytics.log(e.getMessage());
+                    callbackContext.error(e.getMessage());
+                }
+            }
+        });
+    }
+
+    public void deleteChannel(final CallbackContext callbackContext, final String channelID) {
+        cordova.getThreadPool().execute(new Runnable() {
+            public void run() {
+                try {
+                    deleteChannel(channelID);
+                    callbackContext.success();
+                } catch (Exception e) {
+                    Crashlytics.log(e.getMessage());
+                    callbackContext.error(e.getMessage());
+                }
+            }
+        });
+    }
+
+    protected static void deleteChannel(final String channelID){
+        // only call on Android O and above
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            NotificationManager nm = (NotificationManager) cordovaActivity.getSystemService(Context.NOTIFICATION_SERVICE);
+            nm.deleteNotificationChannel(channelID);
+        }
+    }
+
+    public void listChannels(final CallbackContext callbackContext) {
+        cordova.getThreadPool().execute(new Runnable() {
+            public void run() {
+                try {
+                    List<NotificationChannel> notificationChannels = listChannels();
+                    JSONArray channels = new JSONArray();
+                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+                        for (NotificationChannel notificationChannel : notificationChannels) {
+                            JSONObject channel = new JSONObject();
+                            channel.put("id", notificationChannel.getId());
+                            channel.put("name", notificationChannel.getName());
+                            channels.put(channel);
+                        }
+                    }
+                    callbackContext.success(channels);
+                } catch (Exception e) {
+                    Crashlytics.log(e.getMessage());
+                    callbackContext.error(e.getMessage());
+                }
+            }
+        });
+    }
+
+    public static List<NotificationChannel> listChannels(){
+        List<NotificationChannel> notificationChannels = null;
+        // only call on Android O and above
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            NotificationManager nm = (NotificationManager) cordovaActivity.getSystemService(Context.NOTIFICATION_SERVICE);
+            notificationChannels = nm.getNotificationChannels();
+        }
+        return notificationChannels;
+    }
+
+    public static boolean channelExists(String channelId){
+        boolean exists = false;
+        if(Build.VERSION.SDK_INT >= Build.VERSION_CODES.O){
+            List<NotificationChannel> notificationChannels = FirebasePlugin.listChannels();
+            if(notificationChannels != null){
+                for (NotificationChannel notificationChannel : notificationChannels) {
+                    if(notificationChannel.getId() == channelId){
+                        exists = true;
+                    }
+                }
+            }
+        }
+        return exists;
     }
 }